Re-factored test case for Urdu localization.
